# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/articles": {
      "get": {
        "operationId": "ArticlesController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 12,
              "example": 12,
              "type": "number"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "default": "en",
              "example": "ar",
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug (in ?locale=).",
            "schema": {
              "example": "engineering",
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Tag slug (in ?locale=).",
            "schema": {
              "example": "nestjs",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text match on title/excerpt.",
            "schema": {
              "example": "monolith",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated published articles.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PublicArticleListItemEntity"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown or disabled locale.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Malformed query parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "List published articles, resolved and filtered by locale.",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/articles/{slug}": {
      "get": {
        "operationId": "ArticlesController_getBySlug_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "designing-a-modular-monolith",
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Two-letter locale code, validated against enabled locales.",
            "schema": {
              "default": "en",
              "example": "ar",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PublicArticleDetailEntity"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown or disabled locale.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "No published article with this slug in the locale.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Malformed query parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Get a published article by its per-locale slug (drafts 404).",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/admin/articles": {
      "get": {
        "operationId": "ArticlesAdminController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 12,
              "example": 12,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "SCHEDULED",
                "PUBLISHED",
                "ARCHIVED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminArticleEntity"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List articles (any status) with full translation maps.",
        "tags": [
          "articles"
        ]
      },
      "post": {
        "operationId": "ArticlesAdminController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminArticleEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error, slug collision, or invalid publishAt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an article with translations, tags, and scheduling.",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/v1/admin/articles/{id}": {
      "get": {
        "operationId": "ArticlesAdminController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminArticleEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Article not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one article with its full translation map.",
        "tags": [
          "articles"
        ]
      },
      "patch": {
        "operationId": "ArticlesAdminController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminArticleEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Article not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error, slug collision, or invalid publishAt.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an article; upserts translations and status transitions.",
        "tags": [
          "articles"
        ]
      },
      "delete": {
        "operationId": "ArticlesAdminController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Article deleted."
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Article not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an article.",
        "tags": [
          "articles"
        ]
      }
    },
...
  },
  "info": {
    "title": "eslammuatamed API",
    "description": "REST contract for the eslammuatamed platform. Public reads are unauthenticated and locale-resolved; the /admin surface requires a bearer access token (doc 10, doc 19).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "translations[0].slug",
            "description": "Dotted path to the offending field."
          },
          "message": {
            "type": "string",
            "example": "Slug already exists for locale 'en'."
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "/problems/validation",
            "description": "A URI reference identifying the problem type."
          },
          "title": {
            "type": "string",
            "example": "Validation failed"
          },
          "status": {
            "type": "number",
            "example": 422
          },
          "detail": {
            "type": "string",
            "example": "2 fields failed validation."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/admin/articles"
          },
          "errors": {
            "description": "Present on 422 responses only.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "ArticleTaxonomyRefEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Engineering"
          },
          "slug": {
            "type": "string",
            "example": "engineering"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "PublicArticleListItemEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Designing a modular monolith"
          },
          "slug": {
            "type": "string",
            "example": "designing-a-modular-monolith"
          },
          "excerpt": {
            "type": "string",
            "example": "Why one deployable with hard module seams beats microservices here."
          },
          "readingTimeMin": {
            "type": "number",
            "example": 7,
            "description": "Computed reading time in minutes."
          },
          "publishAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "coverImageId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Cover MediaAsset id."
          },
          "category": {
            "$ref": "#/components/schemas/ArticleTaxonomyRefEntity"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleTaxonomyRefEntity"
            }
          },
          "availableLocales": {
            "example": [
              "en",
              "ar"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "excerpt",
          "readingTimeMin",
          "publishAt",
          "coverImageId",
          "category",
          "tags",
          "availableLocales"
        ]
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page (1-based)."
          },
          "perPage": {
            "type": "number",
            "example": 12,
            "description": "Items per page (max 50)."
          },
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total matching items."
          },
          "totalPages": {
            "type": "number",
            "example": 4,
            "description": "Total pages for the current perPage."
          }
        },
        "required": [
          "page",
          "perPage",
          "total",
          "totalPages"
        ]
      },
      "PublicArticleDetailEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Designing a modular monolith"
          },
          "slug": {
            "type": "string",
            "example": "designing-a-modular-monolith"
          },
          "excerpt": {
            "type": "string",
            "example": "Why one deployable with hard module seams beats microservices here."
          },
          "readingTimeMin": {
            "type": "number",
            "example": 7,
            "description": "Computed reading time in minutes."
          },
          "publishAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "coverImageId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Cover MediaAsset id."
          },
          "category": {
            "$ref": "#/components/schemas/ArticleTaxonomyRefEntity"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleTaxonomyRefEntity"
            }
          },
          "availableLocales": {
            "example": [
              "en",
              "ar"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "body": {
            "type": "string",
            "example": "# Heading\n\nOpaque Markdown body…",
            "description": "Opaque Markdown (D01-5)."
          },
          "metaTitle": {
            "type": "string",
            "nullable": true
          },
          "metaDescription": {
            "type": "string",
            "nullable": true
          },
          "ogImageId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "canonicalUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "excerpt",
          "readingTimeMin",
          "publishAt",
          "coverImageId",
          "category",
          "tags",
          "availableLocales",
          "body",
          "metaTitle",
          "metaDescription",
          "ogImageId",
          "canonicalUrl"
        ]
      },
...
      "AdminArticleEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SCHEDULED",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "example": "DRAFT"
          },
          "publishAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "coverImageId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "tagIds": {
            "description": "Associated tag ids.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "translations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/AdminArticleTranslationEntity"
            },
            "description": "Translation map keyed by locale code."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "publishAt",
          "categoryId",
          "coverImageId",
          "tagIds",
          "translations",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "description": "Required category (FR-CNT-050)."
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SCHEDULED",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "default": "DRAFT"
          },
          "publishAt": {
            "type": "string",
            "format": "date-time",
            "description": "Required and future when SCHEDULED."
          },
          "coverImageId": {
            "type": "string",
            "format": "uuid",
            "description": "Cover MediaAsset id."
          },
          "tagIds": {
            "description": "Tag ids (optional).",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "translations": {
            "description": "At least one locale translation.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleTranslationDto"
            }
          }
        },
        "required": [
          "categoryId",
          "translations"
        ]
      },
      "UpdateArticleDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "SCHEDULED",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "publishAt": {
            "type": "string",
            "format": "date-time"
          },
          "coverImageId": {
            "type": "string",
            "format": "uuid"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleTranslationDto"
            }
          }
        }
      },
...
    }
  }
}